Avoid rebooting on windows when upgrading from 5.3.1
A reboot of the machine is triggered when upgrading
from 5.3.1 to a newer version on Windows, if WSL is
not installed.

The v5.3.1 installer is invoked to uninstall the
existing package as part of the upgrade to the new
version. At that point the v5.3.1 installer triggers
a reboot. It doesn't seem to be a simple way to
avoid it.

The trick is to make the uninstallation of the 5.3.1 fail
and explicitly specify to ignore such errors so that
the reboot is not triggered and the upgrade continues.

References:
- https://github.com/orgs/wixtoolset/discussions/8854
- https://github.com/containers/podman/issues/24735

Signed-off-by: Mario Loriedo <mario.loriedo@gmail.com>